+++ app/src/main/java/io/github/sds100/keymapper/ServiceLocator.kt
@@ -36,6 +36,7 @@ import io.github.sds100.keymapper.system.permissions.AndroidPermissionAdapter
 import io.github.sds100.keymapper.system.permissions.SystemFeatureAdapter
 import io.github.sds100.keymapper.system.phone.PhoneAdapter
 import io.github.sds100.keymapper.system.popup.PopupMessageAdapter
+import io.github.sds100.keymapper.system.power.PowerAdapter
 import io.github.sds100.keymapper.system.root.SuAdapter
 import io.github.sds100.keymapper.system.url.OpenUrlAdapter
 import io.github.sds100.keymapper.system.vibrator.VibratorAdapter
@@ -271,6 +272,10 @@ object ServiceLocator {
 
     fun leanbackAdapter(context: Context): LeanbackAdapter {
         return (context.applicationContext as KeyMapperApp).leanbackAdapter
+    }
+
+    fun powerAdapter(context: Context): PowerAdapter {
+        return (context.applicationContext as KeyMapperApp).powerAdapter
     }
 
     private fun createDatabase(context: Context): AppDatabase {